% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeferredMatrix.R
\docType{class}
\name{DeferredMatrix}
\alias{DeferredMatrix}
\alias{DeferredMatrixSeed}
\title{The DeferredMatrix class}
\description{
This has been deprecated in favor of the \linkS4class{ScaledMatrix} class 
from the \pkg{ScaledMatrix} package - use those constructors instead.
}
