% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/runPCA.R
\docType{methods}
\name{runPCA}
\alias{runPCA}
\alias{runPCA,ANY-method}
\title{Principal components analysis}
\usage{
runPCA(x, ...)

\S4method{runPCA}{ANY}(x, rank, center = TRUE, scale = FALSE,
  get.rotation = TRUE, get.pcs = TRUE, as.lem = FALSE, ...)
}
\arguments{
\item{x}{A numeric matrix-like object with samples as rows and variables as columns.}

\item{...}{For the generic, additional arguments to pass to specific methods upon dispatch.

For the \code{ANY} method, further arguments to pass to \code{\link{runSVD}}.
This includes \code{BSPARAM} to specify the algorithm that should be used, and \code{BPPARAM} to control parallelization.}

\item{rank}{Integer scalar specifying the number of principal components to retain.}

\item{center}{A logical scalar indicating whether columns of \code{x} should be centered before the PCA is performed.
Alternatively, a numeric vector of length \code{ncol(x)} containing the value to subtract from each column of \code{x}.}

\item{scale}{A logical scalar indicating whether columns of \code{x} should be scaled to unit variance before the PCA is performed.
Alternatively, a numeric vector of length \code{ncol(x)} containing the scaling factor for each column of \code{x}.}

\item{get.rotation}{A logical scalar indicating whether rotation vectors should be returned.}

\item{get.pcs}{A logical scalar indicating whether the principal component scores should be returned.}

\item{as.lem}{Logical scalar specifying whether the results should be returned as a \linkS4class{LinearEmbeddingMatrix}.
If \code{TRUE}, any \code{FALSE} values of \code{get.pcs} or \code{get.rotation} are ignored.}
}
\value{
By default, a list is returned containing:
\itemize{
\item \code{sdev}, a numeric vector of length \code{rank} containing the standard deviations of the first \code{rank} principal components.
\item \code{rotation}, a numeric matrix with \code{rank} columns and \code{nrow(x)} rows, containing the first \code{rank} rotation vectors.
This is only returned if \code{get.rotation=TRUE}.
\item \code{x}, a numeric matrix with \code{rank} columns and \code{ncol(x)} rows, containing the scores for the first \code{rank} principal components.
This is only returned if \code{get.pcs=TRUE}.
}

If \code{as.lem=TRUE}, a \linkS4class{LinearEmbeddingMatrix} is returned
containing the contents of \code{rotation} as the \code{\link{featureLoadings}};
the contents of \code{x} as the \code{\link{sampleFactors}};
and a \code{sdev} field in the \code{\link{factorData}}.
}
\description{
Perform a principal components analysis (PCA) on a target matrix with a specified SVD algorithm.
}
\details{
This function simply calls \code{\link{runSVD}} and converts the results into a format similar to that returned by \code{\link{prcomp}}.

The generic is exported to allow other packages to implement their own \code{runPCA} methods for other \code{x} objects, e.g., \pkg{scater} for SingleCellExperiment inputs.
}
\examples{
A <- matrix(rnorm(100000), ncol=20)
str(out <- runPCA(A, rank=10))

# Or, as a LinearEmbeddingMatrix.
(out2 <- runPCA(A, rank=10, as.lem=TRUE))

}
\seealso{
\code{\link{runSVD}} for the underlying SVD function.

\code{?\linkS4class{BiocSingularParam}} for details on the algorithm choices.
}
\author{
Aaron Lun
}
